flask_api/requirements.txt
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
networkx==2.2
numpy==1.15.2
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1
